refactor(handlers): extract navigate.ts handler from intent.ts (Stage B)
3672 tests, A1 zero-diff.
